% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis2_fct_data.R
\name{codesmatchingsummary}
\alias{codesmatchingsummary}
\title{Function for summarizing matching/nonmatcing entries for codes}
\usage{
codesmatchingsummary(data, data2, col1)
}
\arguments{
\item{data}{}

\item{data2}{}

\item{col1}{}
}
\value{
Returns a % summary of matching and non-matching entries in column col1 in data and data2
}
\description{
Function for summarizing matching/nonmatcing entries for codes
}
\examples{
codesmatchingsummary(data, data2, col1)
}
