% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_fct_data.R
\name{getDuplicates}
\alias{getDuplicates}
\title{Function for duplicate rows in column}
\usage{
getDuplicates(data, col)
}
\arguments{
\item{data}{input dataset}

\item{col}{name of column}
}
\value{
Returns duplicate data rows in column col in data
}
\description{
Function for duplicate rows in column
}
\examples{
getDuplicates(data, col)
}
