% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_fct_data.R
\name{getMissing}
\alias{getMissing}
\title{Function for empty rows in column}
\usage{
getMissing(data, col)
}
\arguments{
\item{data}{input dataset}

\item{col}{name of column}
}
\value{
Returns empty rows in column col in data
}
\description{
Function for empty rows in column
}
\examples{
getMissing(data, col)
}
