% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_fct_data.R
\name{getsummarymissing}
\alias{getsummarymissing}
\title{Function for summarizing empty rows for two columns}
\usage{
getsummarymissing(data, col1, col2)
}
\arguments{
\item{data}{input dataset}

\item{col1}{name of column}

\item{col2}{name of column}
}
\value{
Returns a summary of missing values % for col1 and col2
}
\description{
Function for summarizing empty rows for two columns
}
\examples{
getsummarymissing(data, col1, col2)
}
