% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_fct_data.R
\name{getsummaryunique}
\alias{getsummaryunique}
\title{Function for summarizing duplicate rows for two columns}
\usage{
getsummaryunique(data, col1, col2)
}
\arguments{
\item{data}{input dataset}

\item{col1}{name of column}

\item{col2}{name of column}
}
\value{
Returns a summary of unique/duplicate values for col1 and col2
}
\description{
Function for summarizing duplicate rows for two columns
}
\examples{
getsummaryunique(data, col1, col2)
}
