% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis2_fct_data.R
\name{namesmismatch}
\alias{namesmismatch}
\title{Function for summarizing the timeliness of names for the same registry code}
\usage{
namesmismatch(data, data2, col1, col2)
}
\arguments{
\item{data}{}

\item{data2}{}

\item{col1}{}

\item{col2}{}
}
\value{
Returns a summary of % of the names matching/mismatching in data and data2 and same value in col1
}
\description{
Function for summarizing the timeliness of names for the same registry code
}
\examples{
namesmismatch(data, data2, col1, col2)
}
