% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis2_fct_data.R
\name{namesmismatchtotal}
\alias{namesmismatchtotal}
\title{Function for evaluating the timeliness of names for the same registry code}
\usage{
namesmismatchtotal(data, data2, col1, col2)
}
\arguments{
\item{data}{}

\item{data2}{}

\item{col1}{}

\item{col2}{}
}
\value{
Returns result of matching/mismatching in col2 with equal value in col1
}
\description{
Function for evaluating the timeliness of names for the same registry code
}
\examples{
namesmismatchtotal(data, data2, col1, col2)
}
