% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis2_fct_data.R
\name{regcodesmatch}
\alias{regcodesmatch}
\title{Function for evaluating matching/nonmatcing entries for codes}
\usage{
regcodesmatch(data, data2, col1)
}
\arguments{
\item{data}{}

\item{data2}{}

\item{col1}{}
}
\value{
Returns a dataframe of matching records (in column col1) in data and data2
}
\description{
Function for evaluating matching/nonmatcing entries for codes
}
\examples{
regcodesmatch(data, data2, col1)
}
